requirements.txt
Django>=3.2,<4
django_compressor>=3
django-mptt>=0.13
django-pagedown<2
django-registration-redux>=2.10
django-reversion>=3.0.5,<4
django-social-share
django-sortedm2m>=3.1.0
django-impersonate
dmoj-wpadmin @ git+https://github.com/DMOJ/dmoj-wpadmin.git
lxml
Pygments
mistune<2
social-auth-core==4.3.0
social-auth-app-django==5.0.0
pytz
django-statici18n
pika
ua-parser
pyyaml
jinja2
django_jinja>=2.5.0
llist
requests
django-fernet-fields @ git+https://github.com/DMOJ/django-fernet-fields.git
pyotp
qrcode[pil]
jsonfield @ git+https://github.com/DMOJ/jsonfield.git
pymoss
packaging<22
celery
ansi2html @ git+https://github.com/DMOJ/ansi2html.git
sqlparse
lupa
martor @ git+https://github.com/DMOJ/martor.git
netaddr
webauthn<1
bleach[css]
django-admin-sortable2<2
icalendar
# This is a celery dependency whose latest major version is breaking everything.
importlib-metadata<5
djangorestframework
djangorestframework-simplejwt
django-cors-headers
djoser
markdown
mysqlclient==2.1.0
django-redis
websocket-client
wheel
uwsgi
